+++ src/v1/Archetypes/Entities/Entity/mustBeEntity.spec.ts
@@ -31,35 +31,72 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+import { expect } from "chai";
+import { describe } from "mocha";
 
-import { DEFAULT_DATA_PATH } from "../../../ErrorHandling/DataPath/defaults/DEFAULT_DATA_PATH";
-import { isType } from "../../../Operators/isType/isType";
-import type { TypeValidatorOptions } from "../../FunctionTypes/TypeValidator/TypeValidatorOptions";
-import type { Entity } from "./Entity";
-import { validateEntity } from "./validateEntity";
+import { AppError, mustBeEntity } from "@safelytyped/core-types";
 
-/**
- * `isEntity()` is a type guard. It proves whether or not the given `input`
- * implements the {@link Entity} protocol.
- *
- * @public
- * @typeParam ID
- * - The type of the entity's ID property.
- * @typeParam T
- * - The type of the wrapped data.
- * @param input -
- * the data to inspect
- * @returns
- * - `true` if:
- *   - `input` implements {@link Entity},
- *   - and if the `implementsEntity()` method returns `true`
- * - `false` otherwise
- */
-export function isEntity<ID = unknown, T = unknown>(
-    input: unknown,
-    {
-        path = DEFAULT_DATA_PATH
-    }: Partial<TypeValidatorOptions> = {}
-): input is Entity<ID,T> {
-    return isType(validateEntity, input, { path });
+class UnitTestEntity
+{
+    public ID: string = "I identify as typesafe!";
+
+    public implementsEntity() {
+        return true;
+    }
+
+    public valueOf() {
+        return "hello unit test!";
+    }
 }
+
+class UnitTestAlmostEntity
+{
+    public ID: string = "I identify as typesafe!";
+
+    public implementsEntity() {
+        return false;
+    }
+
+    public valueOf() {
+        return "hello unit test!";
+    }
+}
+
+describe("mustBeEntity()", () => {
+    it("returns `input` when given an entity", () => {
+        const inputValue = new UnitTestEntity();
+        const expectedValue = inputValue;
+
+        const actualValue = mustBeEntity(inputValue);
+        expect(actualValue).to.eql(expectedValue);
+    })
+
+    describe("throws an `AppError` otherwise", () => {
+        [
+            undefined,
+            null,
+            [ 1, 2, 3, 4, 5 ],
+            [ new UnitTestEntity() ],
+            true,
+            false,
+            100,
+            100.101,
+            {
+                foo: "bar"
+            },
+            new UnitTestAlmostEntity(),
+            "hello world"
+        ].forEach((inputValue) => {
+            it("rejects example " + JSON.stringify(inputValue), () => {
+                let actualValue: any = false;
+                try {
+                    mustBeEntity(inputValue);
+                } catch (e) {
+                    actualValue = e;
+                }
+
+                expect(actualValue).to.be.instanceOf(AppError);
+            })
+        })
+    })
+});